Split markdown typography controls
